Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain[anthropic]>=1.2.7",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
]
